requirements.txt
requests==2.32.3
boto3==1.34.162
python-dotenv==1.0.1
pydantic==2.8.2
tenacity==8.5.0